feat(zsh/ssh-agent): Don't run inside a container
Use $HOSTNAME as the hint (until something more reliable comes up).

+++ zsh/ssh-agent.zsh
@@ -31,35 +31,39 @@ function find_agent_meta {
   fi
 }
 
-find_agent_meta
+if [[ -z "$HOSTNAME" ]]; then
+  find_agent_meta
 
-# If the agent is not running (pid is zero length string)
-if [[ -z "$agent_pid" ]]; then
-    start_agent
-
-# If the agent is running (pid is non zero)
-else
-    # Connect to the currently running ssh-agent
-
-    # this doesn't work because for some reason the ppid is 1 both when
-    # starting from ~/.profile and when executing as `ssh-agent bash`
-    #agent_ppid="$(ps -ef | grep "ssh-agent" | grep -v "grep" | awk '{print($3)}')"
-    agent_ppid=$((agent_pid - 1))
-    echo "Agent ppid $agent_ppid"
-
-    # and the actual auth socket file name is simply numerically one less than
-    # the actual process id, regardless of what `ps -ef` reports as the ppid
-    agent_sock="$(find /tmp -path "*ssh*" -type s -iname "agent.$agent_ppid" 2>&1 | grep -v 'Permission denied')"
-
-    # Invalidate the agent PID if no credentials are attached
-    export_agent_meta $agent_pid $agent_sock
-    ssh-add -l
-    if [[ "$?" = 2 ]]; then
-      killall -9 ssh-agent
-      agent_pid=""
-      echo "Existing agent had no credentials"
+  # If the agent is not running (pid is zero length string)
+  if [[ -z "$agent_pid" ]]; then
       start_agent
-      find_agent_meta
+
+  # If the agent is running (pid is non zero)
+  else
+      # Connect to the currently running ssh-agent
+
+      # this doesn't work because for some reason the ppid is 1 both when
+      # starting from ~/.profile and when executing as `ssh-agent bash`
+      #agent_ppid="$(ps -ef | grep "ssh-agent" | grep -v "grep" | awk '{print($3)}')"
+      agent_ppid=$((agent_pid - 1))
+      echo "Agent ppid $agent_ppid"
+
+      # and the actual auth socket file name is simply numerically one less than
+      # the actual process id, regardless of what `ps -ef` reports as the ppid
+      agent_sock="$(find /tmp -path "*ssh*" -type s -iname "agent.$agent_ppid" 2>&1 | grep -v 'Permission denied')"
+
+      # Invalidate the agent PID if no credentials are attached
       export_agent_meta $agent_pid $agent_sock
-    fi
+      ssh-add -l
+      if [[ "$?" = 2 ]]; then
+        killall -9 ssh-agent
+        agent_pid=""
+        echo "Existing agent had no credentials"
+        start_agent
+        find_agent_meta
+        export_agent_meta $agent_pid $agent_sock
+      fi
+  fi
+else
+  echo "Possibly inside a docker container. Exiting $0."
 fi
